@@ -244,5 +244,5 @@ setup(
     package_data={
         'celiagg': ['data/*'],
     },
-    python_requires=">=3.7",
+    python_requires=">=3.8",
 )
